requirements.txt
python-telegram-bot
requests
pandas
schedule
python-dotenv
flask